% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infragenus.R
\name{basic_infragenus}
\alias{basic_infragenus}
\title{basic_infragenus is used to merge initial URL for infragenus method.}
\usage{
basic_infragenus(
  name = NULL,
  searchOnlyByFirstSymbols = "",
  order = "name",
  orderDirection = "asc",
  author = "",
  originalSpelling = "",
  yearFrom = "",
  yearTo = "",
  rank = FALSE
)
}
\arguments{
\item{name}{The name or partial name of infragenus taxon.}

\item{searchOnlyByFirstSymbols}{Whether to filter records whose Name starts with name parameter, default no.}

\item{order}{Default is "name". Decide to sort the results by name("name") or year("yearFrom").}

\item{orderDirection}{Default is "asc". Determines whether the results are in positive("asc") or reverse("desc") order.}

\item{author}{Select the record authorized by specific person.}

\item{originalSpelling}{Select the record whose original spelling is given string.}

\item{yearFrom}{Records published after the year of "yearFrom".}

\item{yearTo}{Records published before the year of "yearTo".}

\item{rank}{Select records in one or more specific classification levels.
Default is FALSE, not specifying any classification level means not filtering.
Either character type or numeric type can be used. The supported characters of classification levels are c('Genus', 'Infragenus', 'Subgenus', 'Section', 'Subsection', 'Supersection', 'Series', 'Subseries', 'Superseries', 'Unranked'), and the corresponding numbers are c(70, 47, 48, 49, 50, 51, 52, 53, 54, 55).}
}
\value{
URL
}
\description{
basic_infragenus is used to merge initial URL for infragenus method.
}
\examples{
basic_infragenus_url <- basic_infragenus("ty")

}
