% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infraspecies.R
\name{basic_infraspecies}
\alias{basic_infraspecies}
\title{basic_infraspecies is used to merge initial URL for infraspecies method.}
\usage{
basic_infraspecies(
  name = NULL,
  searchOnlyByFirstSymbols = "",
  order = "name",
  orderDirection = "asc",
  author = "",
  originalSpelling = "",
  yearFrom = "",
  yearTo = "",
  rank = FALSE,
  paleoregion = FALSE
)
}
\arguments{
\item{name}{The name or partial name of infraspecies taxon.}

\item{searchOnlyByFirstSymbols}{Whether to filter records whose Name starts with name parameter, default no.}

\item{order}{Default is "name". Decide to sort the results by name("name") or year("yearFrom").}

\item{orderDirection}{Default is "asc". Determines whether the results are in positive("asc") or reverse("desc") order.}

\item{author}{Select the record authorized by specific person.}

\item{originalSpelling}{Select the record whose original spelling is given string.}

\item{yearFrom}{Records published after the year of "yearFrom".}

\item{yearTo}{Records published before the year of "yearTo".}

\item{rank}{Select records in one or more specific classification levels.
Default is FALSE, not specifying any classification level means not filtering.
Either character type or numeric type can be used. The supported characters of classification levels are c('Subspecies', 'Supervarietas', 'Varietas', 'Subvarietas', 'Forma', 'Subforma', 'Infraspecies', 'Unranked'), and the corresponding numbers are c(35, 36, 37, 38, 39, 40, 41, 57).}

\item{paleoregion}{Select records distributed in one or more paleoregions.
Default is FALSE, not specifying any paleoregion means not filtering.
Either character type or numeric type can be used. The supported characters of paleoregions are c('Africa', 'Africa (East)', 'Africa (Equatorial)', 'Africa (North)', 'Africa (South)', 'Africa (West)', 'Altaida (Altaides)', 'Altaj', 'America', 'America (Caribbean)', 'America (North)', 'America (North - Greenland)', 'America (South)', 'Anatolia', 'Angarida', 'Angarida (Mongolia)', 'Antarctica', 'Arctic', 'Australia', 'Australia (New Zealand)', 'Avalonia', 'Baltica', 'Cathaysia', 'Cathaysia (Kitakamiland)', 'Cathaysia (North)', 'Cathaysia (Sino-Korea)', 'Cathaysia (South)', 'Caucasus', 'China (North)', 'China (South)', 'Chingiz', 'Columbia (Amazonia)', 'Columbia (Australia)', 'Columbia (Baltica)', 'Columbia (East Antarctica)', 'Columbia (Greenland)', 'Columbia (Indostan)', 'Columbia (Kalahari)', 'Columbia (Laurentia)', 'Columbia (North Australia)', 'Columbia (North China)', 'Columbia (South Australia)', 'Columbia (South China)', 'Columbia (West Africa)', 'Columbia (West Australia)', 'Cong}
}
\value{
URL
}
\description{
basic_infraspecies is used to merge initial URL for infraspecies method.
}
\examples{
basic_infraspecies_url <- basic_infraspecies("bil")

}
