% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_graber.R
\name{detail_graber}
\alias{detail_graber}
\title{Get detail information of selected items' URLs that should be prepared by brief_information.}
\usage{
detail_graber(items_url)
}
\arguments{
\item{items_url}{Selected items' URLs.}
}
\value{
A dataframe contains detail information
}
\description{
Get detail information of selected items' URLs that should be prepared by brief_information.
}
\examples{
###no example.

}
