% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supragenus.R
\name{supragenus}
\alias{supragenus}
\title{supragenus() is used to implement retrieval from IFPNI's supragenus list accroding to different filting conditions.}
\usage{
supragenus(
  name = NULL,
  searchOnlyByFirstSymbols = "",
  order = "name",
  orderDirection = "asc",
  author = "",
  originalSpelling = "",
  yearFrom = "",
  yearTo = "",
  rank = FALSE
)
}
\arguments{
\item{name}{The name or partial name of supragenus taxon.}

\item{searchOnlyByFirstSymbols}{Whether to filter records whose Name starts with name parameter, default no.}

\item{order}{Default is "name". Decide to sort the results by name("name") or year("yearFrom").}

\item{orderDirection}{Default is "asc". Determines whether the results are in positive("asc") or reverse("desc") order.}

\item{author}{Select the record authorized by specific person.}

\item{originalSpelling}{Select the record whose original spelling is given string.}

\item{yearFrom}{Records published after the year of "yearFrom".}

\item{yearTo}{Records published before the year of "yearTo".}

\item{rank}{Select records in one or more specific classification levels.
Default is FALSE, not specifying any classification level means not filtering.
Either character type or numeric type can be used. The supported characters of classification levels are c('Regnum', 'Subregnum', 'Phylum', 'Subphylum', 'Superphylum', 'Class', 'Subclass', 'Superclass', 'Infraclass', 'Order', 'Superorder', 'Suborder', 'Infraorder', 'Family', 'Superfamily', 'Subfamily', 'Infrafamily', 'Tribe', 'Supertribe', 'Supersubtribe', 'Subtribe', 'Infratribe', 'Turma', 'Anteturma', 'Suprasubturma', 'Subturma', 'Infraturma', 'Unranked'), and the corresponding numbers are c(2, 3, 5, 6, 7, 9, 10, 11, 12, 66, 14, 15, 16, 18, 19, 20, 21, 23, 24, 25, 26, 27, 67, 29, 30, 31, 32, 59).}
}
\value{
A dataframe with detailed information of result.
}
\description{
supragenus() is used to implement retrieval from IFPNI's supragenus list accroding to different filting conditions.
}
\examples{
# For example: find all taxon names start with "cla"
cla <- supragenus("cla", searchOnlyByFirstSymbols = "on")

✔ There are 2 pages found.
✔ There are 18items found.
Collected 18 items from IFPNI. Costed 00:00:20.5


# Get all items of supragenus
# all_supragenus <- supragenus()
✔ There are 205 pages found.
Collected 205 pages from IFPNI. Elapsed Time: 00:05:31.9
✔ There are 2049items found.
Collected 2049 items from IFPNI. Costed 00:51:41
# see rifpni/data/all_supragenus_info.csv
}
