% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_graber.R
\name{basic_graber}
\alias{basic_graber}
\title{basic_graber returns the information of each items according to retrieved result.}
\usage{
basic_graber(basic_url)
}
\arguments{
\item{basic_url}{should be URL path returned by basic methods.}
}
\value{
A dataframe with detailed information of result.
}
\description{
basic_graber returns the information of each items according to retrieved result.
}
\examples{
# supra_cla <- basic_graber(basic_supragenus("cla",searchOnlyByFirstSymbols = "on"))
# genus_cla <- basic_graber(basic_genus("cla",searchOnlyByFirstSymbols = "on"))
}
