% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_graber.R
\name{brief_information}
\alias{brief_information}
\title{Get brief information return by basic_url, a dataframe contains "Name" and "URL".}
\usage{
brief_information(basic_url, output_file = "")
}
\arguments{
\item{basic_url}{created by basic methods, such as basic_supragenus.}
}
\value{
A dataframe contains "Name" and "URL" fields.
}
\description{
Get brief information return by basic_url, a dataframe contains "Name" and "URL".
}
\examples{
# Build initial URL

}
