% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infragenus.R
\name{infragenus}
\alias{infragenus}
\title{infragenus() is used to implement retrieval from IFPNI's infragenus list accroding to different filting conditions.}
\usage{
infragenus(
  name = NULL,
  searchOnlyByFirstSymbols = "",
  order = "name",
  orderDirection = "asc",
  author = "",
  originalSpelling = "",
  yearFrom = "",
  yearTo = "",
  rank = FALSE
)
}
\arguments{
\item{name}{The name or partial name of infragenus taxon.}

\item{searchOnlyByFirstSymbols}{Whether to filter records whose Name starts with name parameter, default no.}

\item{order}{Default is "name". Decide to sort the results by name("name") or year("yearFrom").}

\item{orderDirection}{Default is "asc". Determines whether the results are in positive("asc") or reverse("desc") order.}

\item{author}{Select the record authorized by specific person.}

\item{originalSpelling}{Select the record whose original spelling is given string.}

\item{yearFrom}{Records published after the year of "yearFrom".}

\item{yearTo}{Records published before the year of "yearTo".}

\item{rank}{Select records in one or more specific classification levels.
Default is FALSE, not specifying any classification level means not filtering.
Either character type or numeric type can be used. The supported characters of classification levels are c('Genus', 'Infragenus', 'Subgenus', 'Section', 'Subsection', 'Supersection', 'Series', 'Subseries', 'Superseries', 'Unranked'), and the corresponding numbers are c(70, 47, 48, 49, 50, 51, 52, 53, 54, 55).}
}
\value{
A dataframe with detailed information of result.
}
\description{
infragenus() is used to implement retrieval from IFPNI's infragenus list accroding to different filting conditions.
}
\examples{
# let's see how many taxa names contain "ty" in the infragenus list on IFPNI.
ty_infragenus <- infragenus("ty")
nrow(ty_infragenus)
# [1] 16
# What are they?
ty_infragenus$Name
 #[1] "Anomozamites Platypterygium"        "Calamites Arthropitys"
 #[3] "Calamites Arthropitys"              "Calamites Stylocalamites"
 #[5] "Calamites Stylocalamites"           "Dictyophyllum Camptopteris"
 #[7] "Filicites Dictyopteris"             "Fucoides Dictyotites"
 #[9] "Phlebopteris Dictyopteris"          "Pinites Pityanthus"
 #[11] "Pinites Pityocladus"                "Pinites Pityolepis"
 #[13] "Pinites Pityophyllum"               "Pinites Pityospermum"
 #[15] "Pissadendron Pitys"                 "Zosterophyllum Platyzosterophyllum"

# see data/data/ty_infragenus.rda for details.

# Let's try the most daring operation.
# If you do not pass a value to any argument, you get a dataframe that stores the details of all the taxa in the infragenus list on IFPNI.
# all_infragenus_20231109 <- infragenus()
# This operation may take more than ten minutes. See data/all_infragenus_20231109.rda for the result.
}
